Refactor distance unit conversions: move to separate module and update imports
